driver/powerdriver: Modify DigitalOutputPowerDriver
This driver sits on top of a DigitalOutput and provides a PowerProtocol.

This changes makes the usage of this driver more clear.
Before this change this driver combined resetting a DUT via a GPIO and
power-cycling via PowerSwitch.

Now this driver only uses a PowerSwitch. If someone intends to combine both
methods this sould be placed inside a Strategy.

This change breaks existing environments using the DigitalOutputPowerDriver.

Signed-off-by: Chris Fiege <chris@tinyhost.de>